convert price to double

+++ src/main/java/me/elleuca/PostalService.java
@@ -1,7 +1,7 @@
 package me.elleuca;
 
 public class PostalService {
-    public boolean isFreeShipment(int price) {
-        return price >= 20;
+    public boolean isFreeShipment(double price) {
+        return price >= 20.00;
     }
 }
